requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
yt-dlp>=2024.7.7
PyNaCl>=1.5.0
ffmpeg-python>=0.2.0
aiohttp>=3.8.0
requests>=2.28.0
gspread>=5.12.0
google-auth>=2.23.0
